+++ packages/terra-graphs-docs/src/example-datasets/dataObjects/Line/datasetRegions7.js
@@ -14,47 +14,47 @@ const data = {
       x: new Date(2016, 0, 1, 1, 5).toISOString(),
       y: 5,
       isCritical: false,
-      region: {
+      regions: [{
         start: 3,
         end: 8,
         color: Carbon.helpers.COLORS.GREY,
-      },
+      }],
     },
     {
       x: new Date(2016, 0, 1, 2, 15).toISOString(),
       y: 11,
-      region: {
+      regions: [{
         start: 4,
         end: 14,
         color: Carbon.helpers.COLORS.GREY,
-      },
+      }],
     },
     {
       x: new Date(2016, 0, 1, 3, 15).toISOString(),
       y: 12,
-      region: {
+      regions: [{
         start: 5,
         end: 15,
         color: Carbon.helpers.COLORS.GREY,
-      },
+      }],
     },
     {
       x: new Date(2016, 0, 1, 4, 15).toISOString(),
       y: 16,
-      region: {
+      regions: [{
         start: 10,
         end: 20,
         color: Carbon.helpers.COLORS.GREY,
-      },
+      }],
     },
     {
       x: new Date(2016, 0, 1, 5, 15).toISOString(),
       y: 17,
-      region: {
+      regions: [{
         start: 10,
         end: 20,
         color: Carbon.helpers.COLORS.GREY,
-      },
+      }],
     },
     {
       x: new Date(2016, 0, 1, 6, 15).toISOString(),
@@ -63,68 +63,68 @@ const data = {
     {
       x: new Date(2016, 0, 1, 7, 0).toISOString(),
       y: 11,
-      region: {
+      regions: [{
         start: 8,
         end: 15,
-      },
+      }],
     },
     {
       x: new Date(2016, 0, 1, 8, 15).toISOString(),
       y: 12,
-      region: {
+      regions: [{
         start: 8,
         end: 15,
-      },
+      }],
     },
     {
       x: new Date(2016, 0, 1, 9, 45).toISOString(),
       y: 16,
-      region: {
+      regions: [{
         start: 10,
         end: 20,
-      },
+      }],
     },
     {
       x: new Date(2016, 0, 1, 12, 15).toISOString(),
       y: 17,
-      region: {
+      regions: [{
         start: 10,
         end: 20,
-      },
+      }],
     },
     {
       x: new Date(2016, 0, 1, 13, 15).toISOString(),
       y: 28,
-      region: {
+      regions: [{
         start: 10,
         end: 20,
-      },
+      }],
       isCritical: true,
     },
     {
       x: new Date(2016, 0, 1, 14, 15).toISOString(),
       y: 12,
-      region: {
+      regions: [{
         start: 8,
         end: 15,
-      },
+      }],
     },
     {
       x: new Date(2016, 0, 1, 19, 45).toISOString(),
       y: 13,
-      region: {
+      regions: [{
         start: 9,
         end: 16,
-      },
+      }],
     },
     {
       x: new Date(2016, 0, 1, 21, 15).toISOString(),
       y: 14,
-      region: {
+      regions: [{
         start: 9,
         end: 16,
         color: Carbon.helpers.COLORS.GREY,
-      },
+      }],
     },
   ],
 };