+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/Regions/Line/MultipleValueRegionsOverlapping.test.jsx
@@ -19,29 +19,35 @@ const exampleData = {
       x: new Date(2016, 0, 1, 1, 5).toISOString(),
       y: 5,
       isCritical: false,
-      regions: [{
-        start: 3,
-        end: 8,
-        color: Carbon.helpers.COLORS.GREY,
-      }],
+      regions: [
+        {
+          start: 3,
+          end: 8,
+          color: Carbon.helpers.COLORS.GREY,
+        },
+      ],
     },
     {
       x: new Date(2016, 0, 1, 2, 15).toISOString(),
       y: 11,
-      regions: [{
-        start: 4,
-        end: 14,
-        color: Carbon.helpers.COLORS.GREY,
-      }],
+      regions: [
+        {
+          start: 4,
+          end: 14,
+          color: Carbon.helpers.COLORS.GREY,
+        },
+      ],
     },
     {
       x: new Date(2016, 0, 1, 3, 15).toISOString(),
       y: 12,
-      regions: [{
-        start: 5,
-        end: 15,
-        color: Carbon.helpers.COLORS.GREY,
-      }],
+      regions: [
+        {
+          start: 5,
+          end: 15,
+          color: Carbon.helpers.COLORS.GREY,
+        },
+      ],
     },
     {
       x: new Date(2016, 0, 1, 4, 15).toISOString(),
@@ -50,6 +56,12 @@ const exampleData = {
         start: 10,
         end: 20,
         color: Carbon.helpers.COLORS.GREY,
+
+      },
+      {
+        start: 20,
+        end: 30,
+        color: Carbon.helpers.COLORS.LAVENDER,
       }],
     },
     {
@@ -59,11 +71,21 @@ const exampleData = {
         start: 10,
         end: 20,
         color: Carbon.helpers.COLORS.GREY,
+      },
+      {
+        start: 20,
+        end: 30,
+        color: Carbon.helpers.COLORS.LAVENDER,
       }],
     },
     {
       x: new Date(2016, 0, 1, 6, 15).toISOString(),
       y: 9,
+      regions: [{
+        start: 20,
+        end: 30,
+        color: Carbon.helpers.COLORS.LAVENDER,
+      }],
     },
     {
       x: new Date(2016, 0, 1, 7, 0).toISOString(),
@@ -71,6 +93,11 @@ const exampleData = {
       regions: [{
         start: 8,
         end: 15,
+      },
+      {
+        start: 15,
+        end: 25,
+        color: Carbon.helpers.COLORS.LAVENDER,
       }],
     },
     {
@@ -79,6 +106,11 @@ const exampleData = {
       regions: [{
         start: 8,
         end: 15,
+      },
+      {
+        start: 15,
+        end: 25,
+        color: Carbon.helpers.COLORS.LAVENDER,
       }],
     },
     {
@@ -109,27 +141,43 @@ const exampleData = {
     {
       x: new Date(2016, 0, 1, 14, 15).toISOString(),
       y: 12,
-      regions: [{
-        start: 8,
-        end: 15,
-      }],
+      regions: [
+        {
+          start: 8,
+          end: 15,
+        },
+      ],
     },
     {
       x: new Date(2016, 0, 1, 19, 45).toISOString(),
       y: 13,
-      regions: [{
-        start: 9,
-        end: 16,
-      }],
+      regions: [
+        {
+          start: 9,
+          end: 16,
+        },
+        {
+          start: 20,
+          end: 30,
+          color: Carbon.helpers.COLORS.LAVENDER,
+        },
+      ],
     },
     {
       x: new Date(2016, 0, 1, 21, 15).toISOString(),
       y: 14,
-      regions: [{
-        start: 9,
-        end: 16,
-        color: Carbon.helpers.COLORS.GREY,
-      }],
+      regions: [
+        {
+          start: 9,
+          end: 16,
+          color: Carbon.helpers.COLORS.GREY,
+        },
+        {
+          start: 20,
+          end: 30,
+          color: Carbon.helpers.COLORS.LAVENDER,
+        },
+      ],
     },
   ],
 };